create tests for individual pages
