+++ src/App.js
@@ -1,12 +1,12 @@
 import './App.css';
-import NavHeader from './Components/NavHeader';
+import NavHeader from './Components/NavHeader/NavHeader';
 import ProfileInfoPage from './Pages/ProfileInfoPage';
 import SearchPage from './Pages/SearchPage';
 
 function App() {
   return (
-    <div className="App">
-      <NavHeader isLogoutEnabled={false} isMyAccountEnabled={false} isSearchEnabled={false} />
+    <div className="App" >
+      <NavHeader isLogoutEnabled={false} isSearchEnabled={false} isMyAccountEnabled={false}/>
       <ProfileInfoPage />
       <SearchPage />
     </div>
