+++ src/Pages/SearchPage.test.js
@@ -1,0 +1,8 @@
+import { render, screen } from '@testing-library/react';
+import NavHeader from '../Components/NavHeader/NavHeader';
+
+test('NavHeader.js renders NavHeader', () => {
+    render(<NavHeader />);
+    const testNavbarText = screen.getByTestId('button-div');
+    expect(testNavbarText).toBeInTheDocument();
+});
